requirements.txt
# Smart Rejection - Resume Extraction System
# Core Dependencies

# Supabase for database
supabase>=2.0.0

# PDF Processing
PyPDF2>=3.0.0
pdfplumber>=0.10.0
pdf2image>=1.16.0

# Document Processing (DOCX, DOC)
python-docx>=1.0.0
mammoth>=1.6.0

# Image Processing & OCR
Pillow>=10.0.0
pytesseract>=0.3.10

# NuMarkdown-8B-Thinking (via Hugging Face Transformers)
torch>=2.0.0
transformers>=4.40.0
accelerate>=0.27.0
bitsandbytes>=0.42.0

# Text Processing
python-magic>=0.4.27
chardet>=5.2.0

# Environment & Config
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Utilities
tqdm>=4.66.0
loguru>=0.7.0
uuid>=1.30

# API Framework (optional - for future API development)
fastapi>=0.110.0
uvicorn>=0.27.0
python-multipart>=0.0.9
